int --> size_t (removed -Wsign-compare warning)

+++ ugbase/bridge/misc_bridges/misc_bridge.cpp
@@ -415,7 +415,7 @@ double GetClockS()
 string FilenameStringEscape(string s)
 {
 	stringstream ss;
-	for(int i=0; i<s.length(); i++)
+	for(size_t i=0; i<s.length(); i++)
 	{
 		char c = s[i];
 		if( (c >= '0' && c <= '9' )
